added troubleshooting tips for errors and refactored error code in handlers
